Rename controller
